Removed leaflet-control-geocoder from package.json Removed GeoCoder folder with content from components Removed GeoCoder from GeoLocation Restored MapCenterMarker back to GeoLocation Removed MapCenterMarker.tsx Incorrect wording fix left in SearchForm
