+++ apps/mini-app/package.json
@@ -15,7 +15,6 @@
     "dayjs": "^1.11.10",
     "i18next": "^23.6.0",
     "leaflet": "^1.9.4",
-    "leaflet-control-geocoder": "^2.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.47.0",
