--- apps/mini-app/src/components/GeoCoder/MarkerIcon.tsx
@@ -1,9 +1,0 @@
-import { icon } from 'leaflet';
-
-export default icon({
-  iconSize: [25, 41],
-  iconAnchor: [10, 41],
-  popupAnchor: [2, -40],
-  iconUrl: 'https://unpkg.com/leaflet@1.6/dist/images/marker-icon.png',
-  shadowUrl: 'https://unpkg.com/leaflet@1.6/dist/images/marker-shadow.png',
-});
