--- apps/mini-app/src/components/GeoCoder/index.ts
@@ -1,1 +1,0 @@
-export { GeoCoder } from './GeoCoder';
