+++ apps/mini-app/src/components/GeoLocation/GeoLocation.tsx
@@ -1,10 +1,26 @@
 import { Box } from '@samovar/ui/Box';
-import { MapContainer, TileLayer } from 'react-leaflet';
+import { MapContainer, Marker, TileLayer, useMap, useMapEvents } from 'react-leaflet';
 import type { ComponentProps } from 'react';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import { useMyLocation } from '../../services/ip-geo-location.service';
-import { GeoCoder } from '../GeoCoder';
-import { MapCenterMarker } from './MapCenterMarker';
+
+interface MapCenterMarkerProps {
+  onChange?: (value: [number, number]) => void;
+}
+
+export function MapCenterMarker(props: MapCenterMarkerProps) {
+  const map = useMap();
+  const [position, setPosition] = useState(map.getCenter());
+
+  useMapEvents({
+    click(e) {
+      setPosition(e.latlng);
+      props.onChange?.([e.latlng.lat, e.latlng.lng]);
+    },
+  });
+
+  return <Marker position={position} />;
+}
 
 interface GeoLocationProps {
   value?: [number, number];
@@ -32,7 +48,6 @@ export function GeoLocation({ value, onChange, sx }: GeoLocationProps) {
           url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
         />
         <MapCenterMarker onChange={onChange} />
-        <GeoCoder onChange={onChange} />
       </MapContainer>
     </Box>
   );
